Move floris.tools to top level
