--- floris/tools/__init__.py
@@ -1,50 +1,0 @@
-
-"""
-The :py:obj:`floris.tools` package contains the modules used to drive
-FLORIS simulations and perform studies in various areas of research and
-analysis.
-
-All modules can be imported with
-
-    >>> import floris.tools
-
-The ``__init__.py`` file enables the import of all modules in this
-package so any additional modules should be included there.
-
-Examples:
-    >>> import floris.tools
-
-    >>> dir(floris.tools)
-    ['__builtins__', '__cached__', '__doc__', '__file__', '__loader__',
-    '__name__', '__package__', '__path__', '__spec__', 'cut_plane',
-    'floris_interface',
-    'layout_functions', 'optimization', 'plotting', 'power_rose',
-    'visualization']
-"""
-
-from .floris_interface import FlorisInterface
-from .parallel_computing_interface import ParallelComputingInterface
-from .uncertainty_interface import UncertaintyInterface
-from .visualization import (
-    plot_rotor_values,
-    plot_turbines_with_fi,
-    visualize_cut_plane,
-    visualize_quiver,
-)
-from .wind_data import (
-    TimeSeries,
-    WindRose,
-    WindTIRose,
-)
-
-
-# from floris.tools import (
-#     cut_plane,
-#     floris_interface,
-#     interface_utilities,
-#     layout_functions,
-#     optimization,
-#     plotting,
-#     power_rose,
-#     visualization,
-# )
